feat(plugin-sql-editor): CB-1498 move handler logic to the plugin-sql-editor-navigation-tab
